requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# LangChain for RAG
langchain>=0.1.0
langchain-community>=0.3.27
langchain-openai>=0.0.5

# Vector store
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Embeddings and LLM
openai>=1.10.0
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
